backend/requirements.txt
fastapi==0.105.0
uvicorn==0.23.2
python-multipart==0.0.6
scikit-learn==1.2.2
pandas==2.0.3
numpy==1.24.3
SQLAlchemy==2.0.25
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
joblib==1.3.1
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.7.3
seaborn==0.12.2
google-generativeai==0.3.2 